remove sold item data from prices table. it all lives in the sold items table now

+++ nwmarketapp/views.py
@@ -69,13 +69,6 @@ class PriceSerializer(serializers.ModelSerializer):
             'approved',
             'username',
             'run',
-            'qty',
-            'sold',
-            'gs',
-            'status',
-            'completion_time',
-            'gem',
-            'perk',
         ]
 
 
@@ -126,7 +119,7 @@ class PricesUploadAPI(CreateAPIView):
         ]
 
         serializer = self.get_serializer(data=data, many=True)
-        if not serializer.is_valid():
+        if not serializer.is_valid() and run.section_name != 'Sold Items':
             if run:
                 run.delete()
             return JsonResponse({
